Add validations
Add validations to check connectivity to each component server before executing the script.

+++ Initiate_NW_Connectivity_Checks.ps1
@@ -1,8 +1,9 @@
-$TargetType = ""
+$TargetType = "Windows"
+$TargetPort = "5985"
 $SuccessParams = @{ NoNewLine = $false; ForegroundColor = 'Green' }
 $FailedParams = @{ NoNewLine = $false; ForegroundColor = 'Red' }
 
-
+#Declare CyberArk Component servers
 #$M_CPM_EDN = @("wsapp15260","wsapp15263")
 $M_CPM_EDN = @("wsapp15237","wsapp15260")
 $M_CPM_CMN = @("wsapp15238","wsapp15261")
@@ -37,6 +38,7 @@ $All_P_Hosts = @(($P_CPM_EDN,$P_CPM_CMN,$P_CPM_OPNET),($P_PSM_EDN,$P_PSM_CMN,$P_
 
 $AllHosts = @($All_M_Hosts,$All_P_Hosts)
 
+#Input Validation section
 $EnvInputValidation = {
     try
     {
@@ -60,8 +62,6 @@ $EnvInputValidation = {
         & $EnvInputValidation
     }
 }
-
-
 
 $ComponentInputValidation = {
     try
@@ -111,6 +111,9 @@ $NetworkInputValidation = {
     }
 }
 
+#End Input Validation section
+
+#Dynamically prompt for Inputs that are required to run the script:
 Write-Host "Select the environment`n1. Model`n2. Production";
 $Env = & $EnvInputValidation
 $Env -=1
@@ -119,9 +122,11 @@ Write-Host "Select the Component`n1. CPM`n2. PSM`n3. PSMP";
 $Target_Component = & $ComponentInputValidation
 if($Target_Component -eq 3){
 $TargetType = "Unix"
+$TargetPort = "22"
 }
 else{
 $TargetType = "Windows"
+$TargetPort = "5985"
 }
 $Target_Component -=1
 
@@ -130,50 +135,60 @@ $Target_Network = & $NetworkInputValidation
 $Target_Network -=1
 
 $Script = Read-Host -Prompt 'Specify the script to run:'
+#Input gathhering completed
+
+#Retrieve the component servers based on the inputs:
 $targets = $AllHosts[$Env][$Target_Component][$Target_Network]
+
+#Display selected hosts based on the inputs provided
 if ($targets.count -le 0){
 Write-Host @FailedParams "No Hosts detected with these options.. Exiting..";
 exit
 }
-
-Write-Host @SuccessParams "Selected Targets are: $targets ";
+Write-Host @SuccessParams "Selected component servers are: $targets ";
 
 if($TargetType -eq "Unix"){
 #Write-Host @FailedParams "Unix script is WIP... Exiting.";
 #exit
-#$UnixUser = Read-Host -Prompt 'Provide the Unix user name:'
-$UnixUser = 'e006557'
+#$UnixUser = 'e006557'
+$UnixUser = Read-Host -Prompt 'Provide the Unix user name:'
 
 foreach ($Comp in $targets){
-$DestFile = $Comp+"_Telnet_Result.txt"
-Write-Host "Entering $Comp..";
-$Session = New-PSSession -HostName $Comp -UserName $UnixUser
+$connection = Test-NetConnection -ComputerName $Comp -Port $TargetPort
 
-#Copy the files to the remote host
-Write-Host "Copying the script to $Comp..";
-Copy-Item "hosts.txt" -Destination "\home\$UnixUser\" -ToSession $Session
-Copy-Item $Script -Destination "\home\$UnixUser\" -ToSession $Session
+if ($connection.TcpTestSucceeded) {
+	Write-Host "Connectivity exists to $Comp over Port $TargetPort. Creating a session to the host $Comp";
+	$DestFile = $Comp+"_Telnet_Result.txt"
+	$Session = New-PSSession -HostName $Comp -UserName $UnixUser
 
-Write-Host @SuccessParams "copying the script to $Comp.. IS COMPLETED";
+	#Copy the files to the remote host
+	Write-Host "Copying the script to $Comp..";
+	Copy-Item "hosts.txt" -Destination "\home\$UnixUser\" -ToSession $Session
+	Copy-Item $Script -Destination "\home\$UnixUser\" -ToSession $Session
+	Write-Host @SuccessParams "copying the script to $Comp.. IS COMPLETED";
 
-Write-Host "ENTERING THE SESSION to $Comp..";
-Enter-PSSession $Session
-Write-Host @SuccessParams "Inside the session of $Comp..";
-Invoke-Command -Session $Session {chmod +x check_connectiviti.sh}
-Invoke-Command -Session $Session {./check_connectiviti.sh}
-Exit-PSSession
-Copy-Item "\home\$UnixUser\Telnet_Result.txt" -Destination ".\$DestFile" -FromSession $Session
-Write-Host @SuccessParams "Successfully copied the results from $Comp";
+	Write-Host "ENTERING THE SESSION to $Comp..";
+	Enter-PSSession $Session
+	Write-Host @SuccessParams "Inside the session of $Comp..";
+	Invoke-Command -Session $Session {chmod +x check_connectiviti.sh}
+	Invoke-Command -Session $Session {./check_connectiviti.sh}
+	Exit-PSSession
+	Copy-Item "\home\$UnixUser\Telnet_Result.txt" -Destination ".\$DestFile" -FromSession $Session
+	Write-Host @SuccessParams "Successfully copied the results from $Comp";
 
-#copy the hosts.txt file
-#scp hosts.txt check_connectivity.sh $UnixUser@$Comp:/home/$UnixUser/
-#login to server
-#ssh $UnixUser@$Comp
-#chmod 755 check_connectivity.sh hosts.txt
-#sh check_connectivity.sh
-#exit
-#copy the connectivity_output.txt to server
-#scp $UnixUser@$Comp:/home/$UnixUser/connectivity_output.txt .
+	#copy the hosts.txt file
+	#scp hosts.txt check_connectivity.sh $UnixUser@$Comp:/home/$UnixUser/
+	#login to server
+	#ssh $UnixUser@$Comp
+	#chmod 755 check_connectivity.sh hosts.txt
+	#sh check_connectivity.sh
+	#exit
+	#copy the connectivity_output.txt to server
+	#scp $UnixUser@$Comp:/home/$UnixUser/connectivity_output.txt.
+}
+else {
+	Write-Host @FailedParams "No Connectivity to $Comp over Port $TargetPort. Skipping the Checks from the host $Comp";
+}
 }
 }
 
@@ -183,15 +198,23 @@ $Creds = Get-Credential
 
 Write-Host "Running the script on $targets";
 foreach ($Comp in $targets){
-$DestFile = $Comp+"_Telnet_Result.csv"
-$Session = New-PSSession -ComputerName $Comp -Credential $Creds
-Write-Host "Copying the input file to $Comp..";
-Copy-Item $Script -Destination "C:\Users\$env:username\Documents\" -ToSession $Session
-Copy-Item "Telnet_Input.csv" -Destination "C:\Users\$env:username\Documents\" -ToSession $Session
-Write-Host "Executing script on $Comp..";
-Invoke-Command -ComputerName $Comp -FilePath .\$Script
-Copy-Item "C:\Users\$env:username\Documents\Telnet_Result.csv" -Destination $DestFile -FromSession $Session
-Write-Host "Copied the results from $Comp to this host.";
+$connection = Test-NetConnection -ComputerName $Comp -Port $TargetPort
+
+if ($connection.TcpTestSucceeded) {
+	Write-Host "Connectivity exists to $Comp over Port $TargetPort. Creating a session to the host $Comp";
+	$DestFile = $Comp+"_Telnet_Result.csv"
+	$Session = New-PSSession -ComputerName $Comp -Credential $Creds
+	Write-Host "Copying the input file to $Comp..";
+	Copy-Item $Script -Destination "C:\Users\$env:username\Documents\" -ToSession $Session
+	Copy-Item "Telnet_Input.csv" -Destination "C:\Users\$env:username\Documents\" -ToSession $Session
+	Write-Host "Executing script on $Comp..";
+	Invoke-Command -ComputerName $Comp -FilePath .\$Script
+	Copy-Item "C:\Users\$env:username\Documents\Telnet_Result.csv" -Destination $DestFile -FromSession $Session
+	Write-Host "Copied the results from $Comp to this host.";
+}
+else {
+	Write-Host @FailedParams "No Connectivity to $Comp over Port $TargetPort. Skipping the Checks from the host $Comp";
+}
 }
 }
 Write-Host @SuccessParams "Script execution completed..";
